+++ settings/settings.js
@@ -82,33 +82,33 @@ const config = new DefaultConfig("RiccioFishingUtils", "settings/settings.json")
         setting.redirect("§eChangelog")
     }
 })
-// .addColorPicker({ postponed until amaterasumodded gets updated
-//     category: "RFU",
-//     configName: "bgColor",
-//     subcategory: "Background",
-//     title: "Change Background Color",
-//     description: "Changes the color and alpha of the background",
-//     value: [4, 71, 102, 120],
-// })
-// .addButton({
-//     category: "RFU",
-//     subcategory: "Background",
-//     subcategory: null,
-//     configName: "applyBg",
-//     title: "Apply Changes",
-//     placeHolder: "Apply",
-//     onClick(setting) {
-//         const scheme = JSON.parse(FileLib.read("RiccioFishingUtils", "settings/ColorScheme.json"))
+.addColorPicker({
+    category: "RFU",
+    configName: "bgColor",
+    subcategory: "Background",
+    title: "Change Background Color",
+    description: "Changes the color and alpha of the background",
+    value: [4, 71, 102, 120],
+})
+.addButton({
+    category: "RFU",
+    subcategory: "Background",
+    subcategory: null,
+    configName: "applyBg",
+    title: "Apply Changes",
+    placeHolder: "Apply",
+    onClick(setting) {
+        const scheme = JSON.parse(FileLib.read("RiccioFishingUtils", "settings/ColorScheme.json"))
 
-//         console.log(setting.settings.bgColor)
-//         scheme.Amaterasu.background.color = setting.settings.bgColor
+        console.log(setting.settings.bgColor)
+        scheme.Amaterasu.background.color = setting.settings.bgColor
 
-//         FileLib.write("RiccioFishingUtils", "settings/ColorScheme.json", JSON.stringify(scheme, null, 4))
+        FileLib.write("RiccioFishingUtils", "settings/ColorScheme.json", JSON.stringify(scheme, null, 4))
 
-//         setting.setScheme("settings/ColorScheme.json")
-//         setting.apply()
-//     }
-// })
+        setting.setScheme("settings/ColorScheme.json")
+        setting.apply()
+    }
+})
 .addSwitch({
     category: "General Fishing",
     configName: "seaCreatureCounterToggle",
@@ -1552,6 +1552,14 @@ const setting = new Settings("RiccioFishingUtils", config, "settings/ColorScheme
 })
 .setPos(5, 5)
 .apply()
+
+const scheme = JSON.parse(FileLib.read("RiccioFishingUtils", "settings/ColorScheme.json"))
+if(scheme.Amaterasu.background.color != setting.settings.bgColor) {
+    scheme.Amaterasu.background.color = setting.settings.bgColor
+    FileLib.write("RiccioFishingUtils", "settings/ColorScheme.json", JSON.stringify(scheme, null, 4))
+    setting.setScheme("settings/ColorScheme.json")
+    setting.apply()
+}
 
 register("command", (val) => {
     if(val === 'move') {